requirements.txt
neurodiffeq==0.3.0
PyYAML